+++ trunk/l2jfree-core/src/main/java/com/l2jfree/gameserver/network/clientpackets/RequestPackageSend.java
@@ -60,7 +60,7 @@ public final class RequestPackageSend extends L2GameClientPacket
 		for (int i = 0; i < _count; i++)
 		{
 			int id = readD(); //this is some id sent in PackageSendableList
-			int count = readD();
+			long count = readD();
 			_items.add(new Item(id, count));
 		}
 	}
@@ -102,14 +102,14 @@ public final class RequestPackageSend extends L2GameClientPacket
 				return;
 
 			// Freight price from config or normal price per item slot (30)
-			int fee = _count * Config.ALT_GAME_FREIGHT_PRICE;
-			int currentAdena = player.getAdena();
+			long fee = _count * Config.ALT_GAME_FREIGHT_PRICE;
+			long currentAdena = player.getAdena();
 			int slots = 0;
 
 			for (Item i : _items)
 			{
 				int objectId = i.id;
-				int count = i.count;
+				long count = i.count;
 
 				// Check validity of requested item
 				L2ItemInstance item = player.checkItemManipulation(objectId, count, "deposit");
@@ -152,7 +152,7 @@ public final class RequestPackageSend extends L2GameClientPacket
 			for (Item i : _items)
 			{
 				int objectId = i.id;
-				int count = i.count;
+				long count = i.count;
 
 				// check for an invalid item
 				if (objectId == 0 && count == 0)
@@ -213,9 +213,9 @@ public final class RequestPackageSend extends L2GameClientPacket
 	private class Item
 	{
 		public int	id;
-		public int	count;
+		public long	count;
 
-		public Item(int i, int c)
+		public Item(int i, long c)
 		{
 			id = i;
 			count = c;
